Refactor LED example Wi-Fi helpers into dedicated module
