+++ example/led/main/CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
-    SRCS "main.c" "esp32-lcm.c"
+    SRCS "main.c" "esp32-lcm.c" "esp32-wifi.c"
     REQUIRES freertos esp_wifi esp_event esp_netif nvs_flash driver esp32-homekit esp_timer app_update spi_flash esp_system
 )
