+++ example/led/main/esp32-lcm.h
@@ -5,6 +5,8 @@
 #include <homekit/homekit.h>
 #include <homekit/characteristics.h>
 
+#include "esp32-wifi.h"
+
 #ifndef __HOMEKIT_CUSTOM_CHARACTERISTICS__
 #define __HOMEKIT_CUSTOM_CHARACTERISTICS__
 
@@ -35,13 +37,6 @@ extern "C" {
 
 // Initialiseer NVS en voer automatische herstelactie uit wanneer er geen ruimte is of versie verandert.
 esp_err_t lifecycle_nvs_init(void);
-
-// Start WiFi STA op basis van NVS keys (namespace: wifi_cfg, keys: wifi_ssid, wifi_password).
-// Roep 'on_ready' aan zodra IP is verkregen.
-esp_err_t wifi_start(void (*on_ready)(void));
-
-// Optioneel: stop WiFi netjes
-esp_err_t wifi_stop(void);
 
 // Lifecycle acties die ook door de BOOT-knop of HomeKit aangeroepen worden.
 void lifecycle_request_update_and_reboot(void);
